JPA using more than one entity class

+++ AuthorBoot/src/main/java/com/boot/author/model/Author.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class Author {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long author_id;
+    private int author_id;
     private String name;
     @OneToMany(mappedBy = "author")
     private List<Book> books;
